pyproject.toml
[tool.poetry]
name = "autoreview"
version = "0.1.0"
description = ""
authors = ["Stas Samoylov <zakyska1301@gmail.com>"]
license = "Creative Commons Zero"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
httpx = "^0.27.2"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
openai = "^1.52.1"
requests = "^2.32.3"
aiohttp = "^3.10.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
